[fix] rename directory
