+++ notification-service/src/main/java/com/rainbowgon/notificationservice/domain/kafka/dto/output/MessageOutDto.java
@@ -1,9 +1,9 @@
-package com.rainbowgon.notificationservice.domain.kafka.dto.out;
+package com.rainbowgon.notificationservice.domain.kafka.dto.output;
 
 
-import com.rainbowgon.notificationservice.domain.notification.client.dto.in.BookmarkInDto;
-import com.rainbowgon.notificationservice.domain.notification.client.dto.in.ReservationInDto;
-import com.rainbowgon.notificationservice.domain.notification.client.dto.in.WaitingInDto;
+import com.rainbowgon.notificationservice.domain.notification.client.dto.input.BookmarkInDto;
+import com.rainbowgon.notificationservice.domain.notification.client.dto.input.ReservationInDto;
+import com.rainbowgon.notificationservice.domain.notification.client.dto.input.WaitingInDto;
 import com.rainbowgon.notificationservice.domain.notification.entity.NotificationType;
 import com.rainbowgon.notificationservice.domain.notification.entity.ViewStatus;
 import lombok.*;
