+++ sender-server/src/main/java/com/rainbowgon/senderserver/domain/kafka/dto/input/MessageInDTO.java
